refactor(rs): move cli to its own module
